+++ paper_admin/admin/helpers.py
@@ -16,13 +16,13 @@ from django.forms.forms import DeclarativeFieldsMetaclass
 from django.utils.functional import cached_property
 
 from ..monkey_patch import MonkeyPatchMeta, get_original
-from .widgets import CustomCheckboxInput
+from .widgets import AdminCheckboxInput
 
-checkbox = CustomCheckboxInput({
+checkbox = AdminCheckboxInput({
     "class": "action-select custom-control-input"
 }, lambda value: False)
 
-checkbox_toggle = CustomCheckboxInput({
+checkbox_toggle = AdminCheckboxInput({
     "id": "action-toggle"
 }, lambda value: False)